Exchanged internal long ids for BigInteger. Database interacts via Strings instead.
This should reduce any Trouble we have with to many possible ids, but makes the database access slightly slower.

+++ src/backend/server/src/main/java/prism/core/Property/Probability.java
@@ -21,6 +21,7 @@ import prism.db.mappers.StateAndValueMapper;
 import prism.db.mappers.TransitionMapper;
 import strat.MDStrategy;
 
+import java.math.BigInteger;
 import java.sql.SQLException;
 import java.util.Collections;
 import java.util.Map;
@@ -57,13 +58,13 @@ public class Probability extends Property{
             StateAndValueMapper map = new StateAndValueMapper(project.getModelParser());
 
             vals.iterate(map, false);
-            Map<Long, Double> values = map.output();
+            Map<BigInteger, Double> values = map.output();
 
             project.getDatabase().execute(String.format("ALTER TABLE %s ADD COLUMN %s TEXT", project.getStateTableName(), this.getPropertyCollumn()));
             project.getDatabase().execute(String.format("ALTER TABLE %s ADD COLUMN %s TEXT", project.getTransitionTableName(), this.getPropertyCollumn()));
 
             try (Batch toExecute = project.getDatabase().createBatch(String.format("UPDATE %s SET %s = ? WHERE %s = ?", project.getStateTableName(), this.getPropertyCollumn(), ENTRY_S_ID), 2)) {
-                for (Long stateID : values.keySet()) {
+                for (BigInteger stateID : values.keySet()) {
                     toExecute.addToBatch(String.valueOf(values.get(stateID)), String.valueOf(stateID));
                 }
             } catch (SQLException e) {
