+++ src/backend/server/src/main/java/prism/db/mappers/StateAndValueMapper.java
@@ -3,6 +3,7 @@ package prism.db.mappers;
 import prism.StateAndValueConsumer;
 import prism.core.ModelParser;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
@@ -11,7 +12,7 @@ import java.util.stream.Collectors;
 public class StateAndValueMapper implements StateAndValueConsumer {
 
     private final ModelParser modelParser;
-    private final Map<Long, Double> valueMap;
+    private final Map<BigInteger, Double> valueMap;
 
     public StateAndValueMapper(ModelParser modelParser) {
         this.modelParser = modelParser;
@@ -20,11 +21,11 @@ public class StateAndValueMapper implements StateAndValueConsumer {
 
     @Override
     public void accept(int[] varValues, double value, long stateIndex) {
-        long s_id = modelParser.stateIdentifier(varValues);
+        BigInteger s_id = modelParser.stateIdentifier(varValues);
         valueMap.put(s_id, value);
     }
 
-    public Map<Long, Double> output(){
+    public Map<BigInteger, Double> output(){
         return valueMap;
     }
 }
